refactor: split combined module
